fix(diag): LLP64-correct formats where diagnostics print 64-bit values (MW.1)
Windows x64 is LLP64: 'long' is 4 bytes, so '(unsigned long)' casts of
uintptr_t/pointer values silence -Wformat while truncating the printed
value to its low 32 bits, and the one uncast site is a real varargs
mismatch. All four sites are diagnostics whose whole purpose is the
pointer/word value they print:

  - model.c:  '%lu' fed a raw uintptr_t expression (the sole -Wformat hit
    in the MinGW census) -> '%u' + (unsigned), the value is align&0x3;
  - gfx_pc.c [PC_UNKNOWN_OP], title.c [GUNBARREL-CMD]: Gfx words.w0/w1 are
    uintptr_t (host pointers in w1) -> '%016llX' + (unsigned long long);
  - bg.c [BG-VIS]: the .next pointer -> '%llx' + (unsigned long long).

Remaining '(unsigned long)' print sites were audited and left alone: they
carry bounded offsets/sizes (ROM offsets, buffer deltas) that fit 32 bits
by construction, or DWORDs from GetLastError() where unsigned long is the
correct type.

No output change on LP64 macOS/Linux. Backlog MW.1 (LLP64 sweep).
Verified: MinGW cross lane compiles clean — the model.c -Wformat warning
is gone, no new warnings.

+++ src/game/model.c
@@ -7295,8 +7295,8 @@ void modelSetAnimFrame(Model* model, f32 frame)
     {
         uintptr_t animaddr = (uintptr_t)model->anim;
         if (animaddr < 0x1000 || (animaddr & 0x3)) {
-            fprintf(stderr, "[ANIM_DBG] bad anim ptr %p (align=%lu) for model %p\n",
-                    (void *)model->anim, animaddr & 0x3, (void *)model);
+            fprintf(stderr, "[ANIM_DBG] bad anim ptr %p (align=%u) for model %p\n",
+                    (void *)model->anim, (unsigned)(animaddr & 0x3), (void *)model);
             return;
         }
         /* Also guard modelConstrainOrWrapAnimFrame accesses */
